unfix

+++ server/controllers/jobs.py
@@ -269,12 +269,14 @@ def delete_job():
 
         jobstatus = job.getStatus()
 
-        printer_object.setStatus("complete")
+        if jobstatus == 'printing': # only change statuses, dont remove from queue 
+            printer_object.setStatus("complete")
+        else: 
             # Delete job from the queue
-        queue.deleteJob(job_id) 
+            queue.deleteJob(job_id) 
 
             # Delete job from the database
-        Job.delete_job(job_id)
+            Job.delete_job(job_id)
 
         return jsonify({"success": True, "message": f"Job with ID {job_id} deleted successfully."}), 200
